refactor(map): reorganización modular del sistema de generación y conexión de mapas
- Se renombró `map_manager.py` a `map_builder.py` para reflejar mejor su responsabilidad como constructor de mapas.
- Se dividió `map_generator.py` en módulos especializados:
  - `dungeon_generator.py` → contiene la función `generate_dungeon_map`.
  - `map_merger.py` → contiene `merge_handmade_with_generated`, `connect_from_lobby_exit`, `find_exit_from_lobby`.
  - `geometry.py` → contiene funciones espaciales reutilizables como `intersect`, `center_of`, `create_horizontal_tunnel`, `create_vertical_tunnel`.
- Se actualizaron todos los imports correspondientes en `game.py`, `map_builder.py` y demás archivos relacionados.
- Se mantuvo compatibilidad con `merge_mode=center_to_center`.
- El sistema queda preparado para escalar con futuros modos de conexión, tiles especiales o múltiples zonas handmade.
