+++ roguelike_project/map/handmade_maps/lobby_map.py
@@ -1,0 +1,35 @@
+
+LOBBY_MAP = [
+    "##################################################################################",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#",
+    "#................................................................................#"
+]
