--- roguelike_project/map/lobby_map.py
@@ -1,35 +1,0 @@
-
-LOBBY_MAP = [
-    "#################################################################################",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#",
-    "................................................................................#"
-]
