+++ roguelike_project/map/map_merger.py
@@ -1,67 +1,8 @@
-# roguelike_project/map/map_generator.py
+
 
 import random
-from roguelike_project.config import DUNGEON_WIDTH, DUNGEON_HEIGHT
 
-def intersect(room1, room2):
-    x1a, y1a, x2a, y2a = room1
-    x1b, y1b, x2b, y2b = room2
-    return (
-        x1a <= x2b and x2a >= x1b and
-        y1a <= y2b and y2a >= y1b
-    )
-
-def center_of(room):
-    x1, y1, x2, y2 = room
-    return (x1 + x2) // 2, (y1 + y2) // 2
-
-def create_horizontal_tunnel(map_, x1, x2, y):
-    print(f"🛠️  Crear pasillo horizontal en Y={y}, de X={x1} a X={x2}")
-    for x in range(min(x1, x2), max(x1, x2) + 1):
-        map_[y][x] = "."
-
-def create_vertical_tunnel(map_, y1, y2, x):
-    print(f"🛠️  Crear pasillo vertical en X={x}, de Y={y1} a Y={y2}")
-    for y in range(min(y1, y2), max(y1, y2) + 1):
-        map_[y][x] = "."
-
-def generate_dungeon_map(width=DUNGEON_WIDTH, height=DUNGEON_HEIGHT, max_rooms=10, room_min_size=5, room_max_size=10, return_rooms=False):
-    print("📦 Generando dungeon procedural...")
-    map_ = [["#" for _ in range(width)] for _ in range(height)]
-    rooms = []
-
-    for i in range(max_rooms):
-        w = random.randint(room_min_size, room_max_size)
-        h = random.randint(room_min_size, room_max_size)
-        x = random.randint(1, width - w - 1)
-        y = random.randint(1, height - h - 1)
-
-        new_room = (x, y, x + w, y + h)
-        if any(intersect(room, new_room) for room in rooms):
-            print(f"⛔ Habitación {i} descartada por colisión")
-            continue
-
-        print(f"✅ Habitación {i} aceptada en: {new_room}")
-        for i_ in range(y, y + h):
-            for j in range(x, x + w):
-                map_[i_][j] = "."
-
-        if rooms:
-            prev_x, prev_y = center_of(rooms[-1])
-            new_x, new_y = center_of(new_room)
-            if random.random() < 0.5:
-                create_horizontal_tunnel(map_, prev_x, new_x, prev_y)
-                create_vertical_tunnel(map_, prev_y, new_y, new_x)
-            else:
-                create_vertical_tunnel(map_, prev_y, new_y, prev_x)
-                create_horizontal_tunnel(map_, prev_x, new_x, new_y)
-
-        rooms.append(new_room)
-
-    print(f"🏰 Total habitaciones generadas: {len(rooms)}")
-    if return_rooms:
-        return map_, rooms
-    return map_
+from roguelike_project.map.dungeon_generator import find_closest_room_center, create_horizontal_tunnel, create_vertical_tunnel
 
 def merge_handmade_with_generated(handmade_map, generated_map, offset_x=0, offset_y=0, merge_mode="center_to_center", dungeon_rooms=None):
     print("🔀 Iniciando merge del lobby con dungeon...")
@@ -109,18 +50,6 @@ def connect_from_lobby_exit(map_, lobby_map, offset_x, offset_y, dungeon_rooms):
         create_vertical_tunnel(map_, exit_y, target_y, exit_x)
         create_horizontal_tunnel(map_, exit_x, target_x, target_y)
 
-def find_closest_room_center(source_x, source_y, dungeon_rooms):
-    print(f"🔍 Buscando sala más cercana desde ({source_x}, {source_y})")
-    min_dist = float("inf")
-    closest_center = None
-    for i, room in enumerate(dungeon_rooms):
-        cx, cy = center_of(room)
-        dist = abs(cx - source_x) + abs(cy - source_y)
-        print(f"  🧭 Sala {i}: centro=({cx},{cy}), dist={dist}")
-        if dist < min_dist:
-            min_dist = dist
-            closest_center = (cx, cy)
-    return closest_center
 
 def find_exit_from_lobby(lobby_map, offset_x, offset_y):
     print("🔍 Buscando un '.' en el borde inferior del lobby...")
